Safari-609.1.11.1/LayoutTests/accessibility/mac/loaded-notification-expected.txt

This tests that when enable enhanced accessibility is true and no other accessibility API is called, the load notification will successfully fire.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
